# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Cinnabar API
  description: >-
    This is a Movies/Tv-series Reviews Server based on the OpenAPI 3.0
    specification. 
     If you want to know more about the project here are some useful links:
    - [The Tooren Romaros
    repository](https://github.com/CoPhos/ToorenRomaros/tree/master)

    - [The source API definition for the Tooren Romaros
    project](https://github.com/CoPhos/ToorenRomaros/blob/develop/tooren_romaros_api/src/main/resources/api/openapi.yaml)
  contact:
    name: Jorge Cobos
    url: https://www.JorgeCobos.com
    email: cobos1932@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: http://localhost:9090
    description: Development
  - url: http://localhost:9090
    description: Production
...
paths:
...
  /api/v1/users/images:
    post:
      tags:
        - Image
        - User
        - post
      summary: Create Image for User
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific User. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - CRITIC
            - USER
...
  /api/v1/streamsites/images:
    post:
      tags:
        - Image
        - Stream Site
        - post
      summary: Create Image for Stream Site
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific Stream Site. Users can upload an image file, specify the image
        type, and provide the film ID.
      operationId: createImage_1
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/staffs/images:
    post:
      tags:
        - Image
        - Staff
        - post
      summary: Create Image for Staff
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific staff. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_2
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/posts:
    get:
      tags:
        - Post
        - get
      summary: Get Posts by Custom Query
      description: >-
        This endpoint retrieves posts based on custom query parameters, allowing
        users to filter and sort posts based on specific criteria.
      operationId: getPostByCustomQuery
      parameters:
        - name: tags
          in: query
          description: List of tag IDs to filter posts by.
          required: false
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: isReview
          in: query
          description: Filter posts that are reviews. Default is "false".
          required: false
          schema:
            type: boolean
        - name: latest
          in: query
          description: Sort posts by the latest. Default is "false".
          required: false
          schema:
            type: boolean
        - name: popular
          in: query
          description: Sort posts by popularity. Default is "false"
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          description: Perform a search query to filter posts by title.
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: The number page of the request.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: The page size of each request. Default is "6"
          required: false
          schema:
            type: integer
            format: int32
            default: 6
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostDetailsDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    post:
      tags:
        - post
        - Post
      summary: Create Post
      description: >-
        This endpoint allows users to create a new post by providing the
        necessary details using the provided payload.
      operationId: createPost
      parameters:
        - name: CreatePostDto
          in: query
          description: CreatePostDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
            - CRITIC
  /api/v1/posts/images:
    post:
      tags:
        - Image
        - Post
        - post
      summary: Create Image for Post
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific Post. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_3
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - CRITIC
...
  /api/v1/film/images:
    post:
      tags:
        - Image
        - Film
        - post
      summary: Create Image for Film
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific film. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_4
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: filmId
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/posts/{postId}:
    get:
      tags:
        - Post
        - get
      summary: Get Post by ID
      description: >-
        This endpoint retrieves a specific post identified by its unique
        identifier (postId).
      operationId: getPostById
      parameters:
        - name: postId
          in: path
          description: ID of the Post
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    delete:
      tags:
        - Post
        - delete
      summary: Delete Post
      description: >-
        This endpoint allows users to delete a specific post identified by its
        unique identifier (id). Deleting a post is a permanent action and cannot
        be undone.
      operationId: deletePost
      parameters:
        - name: postId
          in: path
          description: The ID of the Post
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
            - CRITIC
    patch:
      tags:
        - patch
        - Post
      summary: Update Post
      description: >-
        This endpoint allows users to update a specific post identified by its
        unique identifier (id). Users can modify the properties of the post
        using the provided update payload.
      operationId: updatePost
      parameters:
        - name: postId
          in: path
          description: ID of the Post
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: UpdatePostDto
          in: query
          description: UpdatePostDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePostDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - Bearer Authentication:
            - ADMIN
            - CRITIC
...
  /api/v1/{ownerId}/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images by Image Type and Owner ID
      description: >-
        This endpoint retrieves all images associated with a specific entity
        (ownerId) and a specified image type.
      operationId: getAllImagesIdByImageTypeAndOwnerId
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: ownerId
          in: path
          description: >-
            The unique identifier of the entity (e.g., user, film) owning the
            images.
          required: true
          schema:
            type: string
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/users/{userId}/posts:
    get:
      tags:
        - Post
        - get
      summary: Get All Review Post by User ID. Latest post first.
      description: >-
        This endpoint retrieves the all reviews post from a specific user
        (identified by userId).
      operationId: getLatestReviewsByUserIdAndFilmType
      parameters:
        - name: userId
          in: path
          description: ID of the User
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: filmType
          in: query
          description: >-
            Filter films by type (e.g., "all", "1", "2"). Default is "all". (1
            is for movies, 2 is for series).
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: page
          in: query
          description: The number page of the request.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: The page size of each request. Default is "15"
          required: false
          schema:
            type: integer
            format: int32
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostDetailsDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/images/{imageId}:
    get:
      tags:
        - get
        - Image
      summary: Get Image by ID
      description: >-
        This endpoint retrieves an image identified by its unique identifier
        (imageId) from the file system.
      operationId: getImageFromFileSystem
      parameters:
        - name: imageId
          in: path
          description: ID of the Image
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
  /api/v1/films/{imageId}/streamsites/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images from Streamsite by Image Type and Film ID
      description: >-
        This endpoint retrieves all images associated with a streamsite (owner)
        that appear in a specific film (identified by film ID) and filters by a
        specified image type.
      operationId: getAllImagesFromStreamSiteByImageTypeAndStreamSiteId
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/films/{filmId}/staffs/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images from Staff by Image Type and Film ID
      description: >-
        This endpoint retrieves all images associated with staff members
        (owners) that appear in a specific film (identified by film ID) and
        filter by a specified image type.
      operationId: getAllImagesFromStaffByImageTypeAndFilmid
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: filmId
          in: path
          description: The ID of the Film.
          required: true
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        url:
          type: string
        reqMethod:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CreateImageDto:
      required:
        - file
        - filmId
        - filmType
      type: object
      properties:
        file:
          type: string
          format: binary
        filmType:
          type: string
        filmId:
          type: string
...
    CreatePostDto:
      required:
        - film
        - tag
        - user
      type: object
      properties:
        publicationDateTime:
          type: string
          format: date-time
        likeCount:
          minimum: 0
          type: integer
          format: int32
        tittle:
          type: string
        synthesis:
          type: string
        user:
          type: string
        tag:
          type: string
        film:
          type: string
        status:
          type: string
        content:
          type: string
        headline:
          type: string
        review:
          type: boolean
    GetPostDto:
      type: object
      properties:
        id:
          type: string
        publicationDateTime:
          type: string
          format: date-time
        likeCount:
          type: integer
          format: int32
        tittle:
          type: string
        synthesis:
          type: string
        user:
          type: string
        tag:
          type: string
        film:
          type: string
        status:
          type: string
        content:
          type: string
        headline:
          type: string
        mainImageId:
          type: string
        tagId:
          type: string
        review:
          type: boolean
...
    UpdatePostDto:
      type: object
      properties:
        publicationDateTime:
          type: string
          format: date-time
        likeCount:
          minimum: 0
          type: integer
          format: int32
        tittle:
          type: string
        synthesis:
          type: string
        tag:
          type: string
        status:
          type: string
        content:
          type: string
        headline:
          type: string
        review:
          type: boolean
...
    GetListImagesDto:
      type: object
      properties:
        id:
          type: string
        imageSize:
          type: string
        owner:
          type: string
    GetPostDetailsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        publicationDateTime:
          type: string
          format: date-time
        likeCount:
          minimum: 0
          type: integer
          format: int32
        tittle:
          type: string
        synthesis:
          type: string
        headline:
          type: string
        rating:
          type: integer
          format: int32
        user:
          type: string
        tag:
          type: string
        filmMainImageId:
          type: string
        filmId:
          type: string
        mainImageThreeDpi:
          type: string
        mainImageTwoDpi:
          type: string
        mainImageOneDpi:
          type: string
        review:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT